requirements.txt
kornia
pandas
ptflops
plyfile
scipy
tabulate
tifffile
timm
tqdm
wget
diffuser
transformers
open3d
Pillow